Better integrate forking and SSBD verification
When the -e option (expected SSBD bit value) and the -f option (fork
before exec'ing) are both given, a single SSBD bit verification will
occur in the parent prior to forking. The parent will return from
forking and then proceed to verify the SSBD bit value according to the
SECS specified in the -e option's value.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>